+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Caching implementation for Symfony 3 & 4 based on `cache.app` system service
 ### Changed
 - Updating README with Symfony 4, 3, 2 example
 ### Removed